IS-2728: Combine filterpanel to one
